+++ src/main/java/org/crops/fitserver/global/config/SecurityConfig.java
@@ -1,7 +1,6 @@
 package org.crops.fitserver.global.config;
 
 import java.util.Arrays;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.crops.fitserver.domain.user.domain.UserRole;
 import org.crops.fitserver.global.filter.JwtAccessDeniedHandler;
@@ -13,7 +12,6 @@ import org.crops.fitserver.global.security.PrincipalDetailsService;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
@@ -22,9 +20,7 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
 @Configuration
 @EnableWebSecurity
@@ -32,6 +28,7 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 public class SecurityConfig {
 
   private final JwtExceptionFilter jwtExceptionFilter;
+  private final CorsConfigurationSource corsConfigurationSource;
   private final JwtAccessDeniedHandler customAccessDeniedHandler;
   private final HeaderTokenExtractor headerTokenExtractor;
   private final PrincipalDetailsService principalDetailsService;
@@ -75,7 +72,7 @@ public class SecurityConfig {
         )
         .cors(
             cors -> cors
-                .configurationSource(corsConfigurationSource())
+                .configurationSource(corsConfigurationSource)
         )
     ;
 
@@ -97,7 +94,6 @@ public class SecurityConfig {
     return http.build();
   }
 
-  @Bean
   public WebSecurityCustomizer webSecurityCustomizer() {
     return web -> web
         .ignoring()
@@ -110,40 +106,5 @@ public class SecurityConfig {
         .requestMatchers(new AntPathRequestMatcher("/v1/auth/social/**"))
         .requestMatchers(new AntPathRequestMatcher("/h2-console/**")
         );
-  }
-
-  @Bean
-  protected CorsConfigurationSource corsConfigurationSource() {
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    source.registerCorsConfiguration("/**", getDefaultCorsConfiguration());
-    return source;
-  }
-
-  private CorsConfiguration getDefaultCorsConfiguration() {
-    CorsConfiguration configuration = new CorsConfiguration();
-
-    //active profile에 따라 다르게 설정
-    // prod를 제외한 나머지는 모두 localhost:3000을 허용
-
-    var activeProfile = System.getProperty("spring.profiles.active");
-
-    if("prod".equals(activeProfile)) {
-      configuration.setAllowedOrigins(List.of("https://api.f-it.com"));
-    } else {
-      configuration.setAllowedOrigins(List.of("https://api.f-it.com", "http://localhost:3000"));
-    }
-
-    configuration.setAllowedMethods(
-        Arrays.asList(
-            HttpMethod.GET.name(),
-            HttpMethod.POST.name(),
-            HttpMethod.PUT.name(),
-            HttpMethod.PATCH.name(),
-            HttpMethod.DELETE.name()));
-    configuration.setAllowedHeaders(Arrays.asList("*"));
-    configuration.setExposedHeaders(Arrays.asList("Refresh-Token"));
-    configuration.setAllowCredentials(true);
-    configuration.setMaxAge(3600L);
-    return configuration;
   }
 }
